Removed music and added DOCS.md and more
I added a saveRecoveryState() function to try to recover the state in case of a crash.
I removed onclick and used data-action
I updated according to the project with handleProjectAction
I modified the prompt to be more specialized
I removed the music player for now to make the app work, due to a user-agent issue with the streaming service and Tauri.